refactor: 디렉토리 구조 정리
- 컴포넌트 디렉토리 구조 정리
- 페이지 > 컴포넌트 > index.tsx, style.ts(스타일이 긴 경우)
- Header, Footer -> /common
