+++ client/components/Auction/AboutArtist/index.tsx
