+++ client/components/Auction/AuctionList.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 
 import { Button, Center } from '@styles/common';
 import { AuctionCardProps } from '@const/card-type';
-import Card from '@components/Card';
+import Card from '@components/common/Card';
 import { Filter } from '@components/Exhibition/style';
 import { getAuctions } from '@utils/networking';
 import RequireLoginModal from '@components/common/RequireLoginModal';
