+++ client/components/Auction/TrendHistory/index.tsx
