+++ client/components/Exhibition/EditorPage/ColorPicker/index.tsx
