+++ client/components/Exhibition/EditorPage/ImageSlider/index.tsx
