+++ client/components/Exhibition/LabelInput/index.tsx
