+++ client/components/Login/SocialLoginButton/index.tsx
@@ -9,13 +9,22 @@ interface SocialLoginButtonProp {
 
 const SocialLoginButton = ({ type, icon }: SocialLoginButtonProp) => {
     return (
-        <a href={type === 'Google' ? process.env.GOOGLE_AUTH_URL : process.env.KAKAO_AUTH_URL} target='_self'>
+        <a
+            href={
+                type === 'Google'
+                    ? process.env.GOOGLE_AUTH_URL
+                    : process.env.KAKAO_AUTH_URL
+            }
+            target="_self"
+        >
             <Container type={type}>
                 <div>
                     <Image src={icon} alt={`${type} 로그인`} />
                 </div>
                 <Text>
-                    <span style={{ color: type === 'Google' ? 'white' : 'black' }}>
+                    <span
+                        style={{ color: type === 'Google' ? 'white' : 'black' }}
+                    >
                         Sign in with {type}
                     </span>
                 </Text>
